# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/applications": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "获取应用列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Application"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的应用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application"
                ],
                "summary": "创建应用",
                "parameters": [
                    {
                        "description": "Application Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Application"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.CreateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/applications/{id}": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "获取应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Application"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Application"
                ],
                "summary": "更新应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Application Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Application"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Application"
                ],
                "summary": "删除应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/applications/{id}/active_manifest": {
            "patch": {
                "tags": [
                    "Application"
                ],
                "summary": "更新应用的 Active Manifest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Active Manifest Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_api.UpdateActiveManifestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/configurations": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "获取配置列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Configuration"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "创建配置",
                "parameters": [
                    {
                        "description": "Configuration Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Configuration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.CreateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/configurations/{id}": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Configuration"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "更新配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuration Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Configuration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bsonger_devflow-app-service_pkg_model.Application": {
            "type": "object",
            "properties": {
                "active_manifest_id": {
                    "type": "string"
                },
                "active_manifest_name": {
                    "type": "string"
                },
                "config_maps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ConfigMap"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "envs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.EnvVar"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "internet": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Internet"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "replica": {
                    "type": "integer"
                },
                "repo_url": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Service"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ReleaseType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ConfigMap": {
            "type": "object",
            "properties": {
                "files_path": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mount_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.Configuration": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.File"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.EnvVar": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.File": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.Internet": {
            "type": "string",
            "enum": [
                "internal",
                "external"
            ],
            "x-enum-varnames": [
                "Internal",
                "External"
            ]
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.Port": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "target_port": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.ReleaseType": {
            "type": "string",
            "enum": [
                "normal",
                "canary",
                "blue-green"
            ],
            "x-enum-varnames": [
                "Normal",
                "Canary",
                "BlueGreen"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.Service": {
            "type": "object",
            "properties": {
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Port"
                    }
                }
            }
        },
...
        "pkg_api.CreateResponse": {
            "type": "object",
            "properties": {
                "execution_intent_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pkg_api.UpdateActiveManifestRequest": {
            "type": "object",
            "required": [
                "manifest_id"
            ],
            "properties": {
                "manifest_id": {
                    "type": "string"
                }
            }
        },
...
    }
}